Replace JS applied CSS with CSS class
When I first created this script I didn't know that it would make a lot more sense to use CSS classes for style changes instead of styles applied through the script. I replaced the rotate styling with a rot
ate-90 class that rotates the arrow in the sidebar on click.

Signed-off-by: Patrick Kontschak <patrick.kontschak@gmail.com>