Update LESS mixins
I noticed that not only border-radius was used so often without mixins but also box-shadow. New clases are .box-shadow and .inner-shadow (for inset shadow). There are still a few box-shadows with both normal and inset shadow. I have to think about separating them or making a fused class e.g. .full-shadow or something like this.

Signed-off-by: Patrick Kontschak <patrick.kontschak@gmail.com>